# storage: only add a single span to the command queue per batch