--- a/storage/replica.go
+++ b/storage/replica.go
@@ -876,10 +876,10 @@ func (r *Replica) checkBatchRequest(ba roachpb.BatchRequest) error {
 // already in the queue. Returns a cleanup function to be called when the
 // commands are done and can be removed from the queue.
 func (r *Replica) beginCmds(ba *roachpb.BatchRequest) func(*roachpb.BatchResponse, *roachpb.Error) {
-	var cmdKeys []interface{}
+	var cmdEntry interface{}
 	// Don't use the command queue for inconsistent reads.
 	if ba.ReadConsistency != roachpb.INCONSISTENT {
-		var spans []roachpb.Span
+		spans := make([]roachpb.Span, 0, len(ba.Requests))
 		readOnly := ba.IsReadOnly()
 		for _, union := range ba.Requests {
 			h := union.GetInner().Header()
@@ -888,7 +888,7 @@ func (r *Replica) beginCmds(ba *roachpb.BatchRequest) func(*roachpb.BatchRespons
 		var wg sync.WaitGroup
 		r.mu.Lock()
 		r.mu.cmdQ.GetWait(readOnly, &wg, spans...)
-		cmdKeys = append(cmdKeys, r.mu.cmdQ.Add(readOnly, spans...)...)
+		cmdEntry = r.mu.cmdQ.Add(readOnly, spans...)
 		r.mu.Unlock()
 		wg.Wait()
 	}
@@ -911,13 +911,13 @@ func (r *Replica) beginCmds(ba *roachpb.BatchRequest) func(*roachpb.BatchRespons
 	}
 
 	return func(br *roachpb.BatchResponse, pErr *roachpb.Error) {
-		r.endCmds(cmdKeys, ba, br, pErr)
+		r.endCmds(cmdEntry, ba, br, pErr)
 	}
 }
 
 // endCmds removes pending commands from the command queue and updates
 // the timestamp cache using the final timestamp of each command.
-func (r *Replica) endCmds(cmdKeys []interface{}, ba *roachpb.BatchRequest, br *roachpb.BatchResponse, pErr *roachpb.Error) {
+func (r *Replica) endCmds(cmdEntry interface{}, ba *roachpb.BatchRequest, br *roachpb.BatchResponse, pErr *roachpb.Error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	// Only update the timestamp cache if the command succeeded and is
@@ -948,7 +948,7 @@ func (r *Replica) endCmds(cmdKeys []interface{}, ba *roachpb.BatchRequest, br *r
 			}
 		}
 	}
-	r.mu.cmdQ.Remove(cmdKeys)
+	r.mu.cmdQ.Remove(cmdEntry)
 }
 
 // applyTimestampCache moves the batch timestamp forward depending on