Previously, the CommandQueue was fine-grained. This switches to a
coarse-grained approach were a single span is computed for a batch that
covers all of the keys used by the individual KV ops. This reduces
parallelism at the range level, but dramatically reduces the
CommandQueue overhead for bulk operations.

```
name                  old time/op    new time/op    delta
KVInsert1_SQL-32         492µs ± 2%     489µs ± 3%     ~     (p=0.072 n=20+20)
KVInsert10_SQL-32        935µs ± 2%     901µs ± 1%   -3.59%  (p=0.000 n=19+19)
KVInsert100_SQL-32      5.09ms ± 2%    4.61ms ± 4%   -9.39%  (p=0.000 n=20+20)
KVInsert1000_SQL-32     49.9ms ± 6%    42.0ms ± 5%  -15.97%  (p=0.000 n=19+20)
KVInsert10000_SQL-32     659ms ±12%     470ms ±10%  -28.63%  (p=0.000 n=19+18)

name                  old allocs/op  new allocs/op  delta
KVInsert1_SQL-32           304 ± 0%       298 ± 0%   -1.97%  (p=0.000 n=16+19)
KVInsert10_SQL-32          754 ± 0%       712 ± 0%   -5.57%  (p=0.000 n=17+17)
KVInsert100_SQL-32       5.04k ± 0%     4.64k ± 0%   -8.00%  (p=0.000 n=19+20)
KVInsert1000_SQL-32      48.7k ± 2%     44.0k ± 1%   -9.69%  (p=0.000 n=20+20)
KVInsert10000_SQL-32      570k ±15%      477k ± 8%  -16.32%  (p=0.000 n=20+20)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6412)
<!-- Reviewable:end -->