@petermattis 2016-04-29
@bdarnell This would be an alternative to #6411. I'm not sure of all of the repercussions of doing this. Perhaps there is a middle ground between the coarse and fine grained approaches, though I'm not seeing what it is right now.