@petermattis 2016-04-29
@tschottdorf Doing something similar to this PR for the timestamp cache yields the following additional improvements:

```
name                  old time/op    new time/op    delta
KVInsert1_SQL-32         498µs ± 3%     499µs ± 3%     ~     (p=0.529 n=20+20)
KVInsert10_SQL-32        914µs ± 2%     888µs ± 2%   -2.89%  (p=0.000 n=20+20)
KVInsert100_SQL-32      4.59ms ± 2%    4.25ms ± 3%   -7.43%  (p=0.000 n=19+20)
KVInsert1000_SQL-32     42.5ms ± 8%    37.8ms ± 6%  -11.15%  (p=0.000 n=20+20)
KVInsert10000_SQL-32     489ms ±10%     449ms ±12%   -8.11%  (p=0.000 n=19+19)
```

Not as dramatic, but still worth doing from a performance perspective. Once again, I don't understand all of the repercussions.