@bdarnell 2016-04-29
There are some workloads for which this coarse-grained approach would be disastrous for performance (the Google Reader schema would have problems with this, for instance). In general, the bigger ranges get, the worse the coarse-grained command queue becomes (and I believe we'll want to increase our range sizes in the future). But it's definitely a big win for bulk inserts. 

Could we come up with a heuristic to switch between the two modes? For example, if we could afford to go to the DB here we could see if there are any keys that exist between the min and max of the batch. (if not, contention is unlikely and coarse mode wins). Or maybe we could cache a sample of keys at regular intervals in the range to judge how much of the range would be covered by a coarse entry in the command queue.

Teaching this layer about our key structure might give us another middle ground: strip the column IDs from the end of structured keys, and turn those from point operations into span operations on the entire row. This would reduce the number of command queue entries in the bulk insert case by a significant factor (but not as much as the fully coarse-grained approach) without causing much extra contention.