@petermattis 2016-04-29
All keys are structured keys now (needed to always split at row boundaries). We could definitely strip off the column IDs so that all of the keys for a row get joined. At a minimum, that seems worthwhile. Do you mind explaining why the Google Reader schema would have problems with this approach? Seems like this approach decreases potential parallelism when you have a request to a table that needs to perform multiple distinct operations (reads or writes) that are spread out, but clustered enough so that they hit a single range. If the read/write operations are contiguous, joining the spans like this will be equivalent. 

There is no rush to get this changed merged. If you can describe a workload you think it will hurt, I'll code up a benchmark so that we can see the before and after.