@bdarnell 2016-04-29
The Reader schema (which was tailored to the properties of bigtable so I'm not sure we'd do it the same way on cockroach) had a table that mapped article IDs to the set of users who had read them (there were other tables that were user-first, but this table was the only one where write contention would really be an issue). So if you marked several articles as read at once (not a common case - it mainly happened when the client was recovering from being offline), and those articles were on the same range (because they were from the same source), you'd block other users from being able to mark those articles as read. 

So for a benchmark, I would make a table like `(article_id int, user_id int, primary key (article_id, user_id))` and insert batches of tuples into it. All tuples in a batch would have the same user_id; article_ids would be chosen from some power-law distribution.