@petermattis 2016-04-29
Added an `InsertDistinct` benchmark. I had to add a unique ID to each row to prevent collisions with the random ID generation. Doesn't seem to be showing any negative to this change. Suggestions welcome.