@bdarnell 2016-04-29
Command queue contention is a more severe problem in multi-node configurations, because commands will remain pending in the queue for at least a network round trip. But that's a good sign that this change was a net positive even with the increased contention.