@petermattis 2016-04-29
I tested inserting a simulated network delay by doing a `time.Sleep` after the `CommandQueue` wait is performed. Here is a 1 millisecond sleep:

```
name                           old time/op  new time/op  delta
InsertDistinct1_Cockroach-8    3.90ms ± 2%  3.67ms ± 3%   -5.88%    (p=0.000 n=9+9)
InsertDistinct10_Cockroach-8   1.52ms ± 3%  2.87ms ± 5%  +88.50%  (p=0.000 n=10+10)
InsertDistinct100_Cockroach-8  2.20ms ± 8%  3.68ms ± 4%  +67.30%   (p=0.000 n=10+9)
```

And here is a 10 millisecond sleep:

```
name                           old time/op  new time/op   delta
InsertDistinct1_Cockroach-8    15.9ms ± 2%   15.6ms ± 3%    -1.95%  (p=0.009 n=10+10)
InsertDistinct10_Cockroach-8   1.84ms ± 3%  14.48ms ± 3%  +687.42%   (p=0.000 n=8+10)
InsertDistinct100_Cockroach-8  2.35ms ± 7%  15.88ms ± 3%  +576.24%   (p=0.000 n=8+10)
```

Clearly, the decrease in parallelism is awful for this workload if we have additional network latencies.