@petermattis 2016-04-29
I wonder if we can get the best of both worlds. When we initially add the spans to the timestamp cache we insert a single covering span. When `CommandQueue.GetWait` detects an overlap, we replace the covering span with the individual spans. This would certainly be more complex, but I think it is doable.