@petermattis 2016-04-30
I've added a new implementation which expands the "covering" command queue entry when a conflict is detected. This allows for the fine-grained parallelism when conflicts are detected, but avoids the overhead of adding all of the individual spans during bulk insert workloads.

When I get time I'm going to investigate the `InsertDistinct100` degradation. It is higher than I expect. 

PTAL.