@petermattis 2016-05-03
@bdarnell Nothing but net. The numbers below are with an artificial 10ms pause to simulate network latency. They differ from the ones I posted above because the benchmarks were run on a different machine:

```
name                            old time/op    new time/op    delta
InsertDistinct1_Cockroach-32      14.0ms ± 3%    13.9ms ± 3%     ~     (p=0.478 n=20+20)
InsertDistinct10_Cockroach-32     2.43ms ± 6%    2.29ms ± 5%   -5.62%  (p=0.000 n=20+19)
InsertDistinct100_Cockroach-32    3.02ms ±13%    2.70ms ± 5%  -10.79%  (p=0.000 n=20+20)

name                            old allocs/op  new allocs/op  delta
InsertDistinct1_Cockroach-32       2.48k ±10%     2.31k ± 9%   -6.57%  (p=0.000 n=20+20)
InsertDistinct10_Cockroach-32      2.47k ± 5%     2.40k ± 3%   -3.10%  (p=0.000 n=20+19)
InsertDistinct100_Cockroach-32     5.20k ± 3%     5.09k ± 1%   -2.13%  (p=0.000 n=17+18)
```